RandomCropGrowth 2
Push 2 cause the first one doesn'T seem to have been pushed properly?
Notes:
Crop growth speed might need a bit of adjustment if it were to be implemented. Speed is like this mainly so that the growth is visible for testing purposes.

Also slight name changes:
Renamed "immature" loot table to "sprout"
Renamed TwoTallCropBlock to TwoTallCrops